change ints to longs to prevent potential overflows

+++ cpp/euler7.cpp
@@ -10,8 +10,8 @@ using namespace std;
 /*
  * Built a list of numbers up to n.
  */
-void range(int n, int nums[]) {
-    for (int i = 0; i < n; i++) {
+void range(long n, long nums[]) {
+    for (long i = 0; i < n; i++) {
         nums[i] = i;
     }
 }
@@ -21,9 +21,9 @@ void range(int n, int nums[]) {
  * Build a list of primes up to n.
  * The list will contain 0s at all non primes indices.
  */
-void primes(int n, int sieve[]) {
-    int root;
-    int index;
+void primes(long n, long sieve[]) {
+    long root;
+    long index;
     range(n, sieve);
     sieve[1] = 0;
     if (n < 2) {
@@ -33,7 +33,7 @@ void primes(int n, int sieve[]) {
     index = 0;
     while (index <= root) {
         if (sieve[index] != 0) {
-            int i = pow(index, 2);
+            long i = pow(index, 2);
             while (i < n) {
                 sieve[i] = 0;
                 i += index;
@@ -44,12 +44,12 @@ void primes(int n, int sieve[]) {
 }
 
 int main() {
-    int target = 10001;
-    int n = 110000;
-    int nums[n];
+    long target = 10001;
+    long n = 110000;
+    long nums[n];
     primes(n, nums);
-    int count = 0;
-    for (int i = 0; i < n; i++) {
+    long count = 0;
+    for (long i = 0; i < n; i++) {
         if (nums[i] != 0) {
             count++;
             if (count == target) {
